refactor: feature 폴더 접근법에 따라 폴더명 변경
