+++ src/redux/itemList/thunk.ts
@@ -2,7 +2,7 @@ import { BASE_URL } from 'apis';
 import axios from 'axios';
 import type { Dispatch } from 'redux';
 
-import { ItemListAction, ItemListActionType } from './itemList';
+import { ItemListAction, ItemListActionType } from './action';
 
 export const getItemList = () => async (dispatch: Dispatch<ItemListAction>) => {
   dispatch({ type: ItemListActionType.GET_ITEM_LIST_START });
