+++ src/redux/rootReducer.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 
-import { cartListReducer } from './cartList/cartListReducer';
-import { itemListReducer } from './itemList/itemListReducer';
+import { cartListReducer } from './cartList/reducer';
+import { itemListReducer } from './itemList/reducer';
 
 const rootReducer = combineReducers({ itemListReducer, cartListReducer });
